requirements.txt
PyQt6>=6.5.0
PyQt6-WebEngine>=6.5.0
numpy>=1.24
Pillow>=10.0
pytest>=7.4
pytest-qt>=4.2
